# specs/001-adherence-history/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Med Manager - Adherence API (Feature 001)
  version: 0.1.0
servers:
  - url: /api

paths:
  /adherence/taken:
    post:
      summary: Create TAKEN adherence event
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema: { type: string, minLength: 1, maxLength: 128 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTakenRequest"
      responses:
        "200":
          description: TAKEN created (or idempotent replay)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdherenceEvent"
        "400":
          description: Invalid argument
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Unauthorized/Expired
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Authorization denied
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "409":
          description: Duplicate
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  /adherence/today:
    get:
      summary: Get today doses with optional adherence event (0/1) + month marks
      parameters:
        - in: query
          name: targetPatientId
          required: false
          schema: { type: string }
        - in: query
          name: date
          required: false
          schema: { type: string, format: date }
      responses:
        "200":
          description: Today payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodayResponse"
        "400":
          description: Invalid argument
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Unauthorized/Expired
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Authorization denied
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  /adherence/history:
    get:
      summary: Get adherence history by date range with cursor pagination
      parameters:
        - in: query
          name: targetPatientId
          required: false
          schema: { type: string }
        - in: query
          name: from
          required: true
          schema: { type: string, format: date }
        - in: query
          name: to
          required: true
          schema: { type: string, format: date }
        - in: query
          name: limit
          required: false
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - in: query
          name: cursor
          required: false
          schema: { type: string }
      responses:
        "200":
          description: History page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryResponse"
        "400":
          description: Invalid argument / invalid cursor / limit exceeded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Unauthorized/Expired
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Authorization denied
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

components:
  schemas:
    CreateTakenRequest:
      type: object
      additionalProperties: false
      properties:
        targetPatientId:
          type: string
          description: family only; patient is implicit
        medicationId:
          type: string
        scheduledAt:
          type: string
          format: date-time
      required: [medicationId, scheduledAt]

    AdherenceEvent:
      type: object
      additionalProperties: false
      properties:
        adherenceEventId: { type: string }
        patientId: { type: string }
        medicationId: { type: string }
        scheduledAt: { type: string, format: date-time }
        status:
          type: string
          enum: [TAKEN, MISSED, RESOLVED]
        recordedBy:
          type: string
          enum: [patient, family]
        recordedByUserId: { type: string }
        createdAt: { type: string, format: date-time }
        resolvedAt: { type: string, format: date-time, nullable: true }
      required:
        [adherenceEventId, patientId, medicationId, scheduledAt, status, recordedBy, recordedByUserId, createdAt]

    Dose:
      type: object
      additionalProperties: false
      properties:
        doseKey: { type: string }
        patientId: { type: string }
        medicationId: { type: string }
        slot:
          type: string
          enum: [morning, noon, night, bed]
        scheduledAt: { type: string, format: date-time }
        event:
          anyOf:
            - $ref: "#/components/schemas/AdherenceEvent"
            - type: "null"
      required: [doseKey, patientId, medicationId, slot, scheduledAt, event]

    MonthMark:
      type: object
      additionalProperties: false
      properties:
        date: { type: string, format: date }
        hasEvents: { type: boolean }
      required: [date, hasEvents]

    TodayResponse:
      type: object
      additionalProperties: false
      properties:
        doses:
          type: array
          items: { $ref: "#/components/schemas/Dose" }
        monthMarks:
          type: array
          items: { $ref: "#/components/schemas/MonthMark" }
      required: [doses, monthMarks]

    HistoryResponse:
      type: object
      additionalProperties: false
      properties:
        events:
          type: array
          items: { $ref: "#/components/schemas/AdherenceEvent" }
        nextCursor:
          type: string
          nullable: true
      required: [events, nextCursor]

    ErrorResponse:
      type: object
      additionalProperties: false
      properties:
        error_code: { type: string }
        message: { type: string }
        details:
          type: object
          additionalProperties: true
          nullable: true
      required: [error_code, message]